# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API do Sistema Financeiro
  description: |
    API completa para gerenciamento de finanças pessoais e empresariais.
    
    ## Funcionalidades Principais
    
    ### Gestão Financeira
    - **Contas**: Gerenciamento de contas bancárias e carteiras
    - **Transações**: Registro de receitas e despesas
    - **Categorias**: Organização de transações por categorias
    - **Relatórios**: Análises e relatórios financeiros
    
    ### Gestão de Clientes e Fornecedores
    - **Clientes**: Cadastro e gestão de clientes
    - **Fornecedores**: Controle de fornecedores
    - **Recebíveis**: Controle de contas a receber
    - **Pagáveis**: Controle de contas a pagar
    
    ### Investimentos e Financiamentos
    - **Investimentos**: Gestão de portfólio de investimentos
    - **Metas de Investimento**: Definição e acompanhamento de metas
    - **Financiamentos**: Controle de empréstimos e financiamentos
    - **Pagamentos**: Gestão de parcelas e pagamentos
    
    ### Recursos Avançados
    - **Autenticação JWT**: Sistema seguro de autenticação
    - **Notificações**: Sistema de alertas e lembretes
    - **Jobs**: Processamento em background
    - **Validações**: Validação robusta de dados
    
    ## Status dos Testes
    - **13/17 suítes de integração** funcionando 100%
    - **142/215 testes** passando (66%)
    - **Execução sequencial** implementada para evitar conflitos
    - **Cobertura de código** configurada
    
    ## Documentação Relacionada
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login
...
security:
  - bearerAuth: []
...
paths:
...
  /api/creditors:
    get:
      tags:
        - Creditors
      summary: Listar credores
      description: Lista todos os credores do usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de credores
    post:
      tags:
        - Creditors
      summary: Criar credor
      description: Cria um novo credor
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Credor criado
  /api/creditors/{id}:
    get:
      tags:
        - Creditors
      summary: Obter credor
      description: Obtém um credor específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do credor
    put:
      tags:
        - Creditors
      summary: Atualizar credor
      description: Atualiza um credor existente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Credor atualizado
    delete:
      tags:
        - Creditors
      summary: Excluir credor
      description: Remove um credor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Credor removido

  # Endpoints de Integridade de Dados (Admin)
...